+++ CHANGELOG.md
@@ -133,3 +133,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.0.0] - 2026-02-17
 
 - Initial release
+## [1.12.0]
+- Added a new command "Pullgod: View Pull Requests for Remote..." to allow viewing PRs from other git remotes.
+- Added a new setting "pullgod.preferredRemote" to configure the default git remote to fetch pull requests from.